feat: :art: Agregando modulo usuarios
Se agrego modulo usuarios, creacion y eliminacion, falta edicion

Usuarios

+++ resources/js/Pages/Users/composables/useIndex.js
@@ -1,5 +1,17 @@
-import { alertConfirm, alertSuccess, validateRequest } from "@/Util/alerts";
-import { success, workConfirmDelete, workDeleted } from "@/Util/messages";
+import {
+    alertConfirm,
+    alertInfo,
+    alertSuccess,
+    validateRequest,
+} from "@/Util/alerts";
+import { user } from "@/Util/auth";
+import {
+    success,
+    userCantDelete,
+    userCantDeleteInSesion,
+    userConfirmDelete,
+    userDeleted,
+} from "@/Util/messages";
 import { useForm, usePage } from "@inertiajs/vue3";
 import { FilterMatchMode } from "primevue/api";
 import { ref, toRefs } from "vue";
@@ -8,15 +20,19 @@ export default function useIndex() {
     const { props } = usePage();
     const form = useForm({});
     const items = ref(props.items || []);
-    const pathInvWork = ref(props.paths?.inv_work || "");
     const loading = ref(false);
-    const dtworks = ref();
+    const dtusers = ref();
     const filters = ref({
         global: { value: null, matchMode: FilterMatchMode.CONTAINS },
     });
 
-    const confirmDeleteItem = (id) => {
-        alertConfirm({ title: workConfirmDelete }).then((resp) => {
+    const confirmDelete = (id) => {
+        if (id === user.value.id) {
+            alertInfo(userCantDelete, userCantDeleteInSesion);
+            return;
+        }
+
+        alertConfirm({ title: userConfirmDelete }).then((resp) => {
             if (resp.isConfirmed) {
                 deleteItem(id);
             }
@@ -25,14 +41,12 @@ export default function useIndex() {
 
     const deleteItem = (id) => {
         loading.value = true;
-        const url = route("investigation-works.destroy", {
-            work: id,
-        });
+        const url = route("users.destroy", { user: id });
 
         form.delete(url, {
             preserveScroll: true,
             onSuccess: () => {
-                alertSuccess(success, workDeleted);
+                alertSuccess(success, userDeleted);
                 globalThis.location.reload();
             },
             onError: (errors) => validateRequest(errors),
@@ -41,18 +55,17 @@ export default function useIndex() {
     };
 
     const exportCSV = () => {
-        dtworks.value.exportCSV();
+        dtusers.value.exportCSV();
     };
 
     return {
-        confirmDeleteItem,
+        confirmDelete,
         exportCSV,
-        dtworks,
+        dtusers,
         ...toRefs({
             filters,
             items,
             loading,
-            pathInvWork,
         }),
     };
 }
